requirements.txt
kivy>=2.2.1
kivymd>=1.1.1
Pillow>=10.1.0

pymysql>=1.0.2
plyer>=2.1.0

python-docx>=0.8.11
xlsxwriter>=3.1.2